Send signed JWT
Create a signed JWT and render it to the page from 'web_server'. This
can then be sent to the gateway (local_server) via the browser extension and the
signature verified to prevent unauthorised messages.

Signature verification will be done in a subsequent commit.

Reorganise the directory structure.
