+++ src/chrome_extension/manifest.json
